refactor/ 테스트를 api 개수당 1개로 쪼갬
